IOS-4855 Moved layout static func logic to separate file
